+++ CHANGES.rst
@@ -51,6 +51,8 @@ Bugfixes
 - Prevent room booking sidebar menu from overlapping with the user dropdown menu
   (:pr:`5910`)
 - Allow cancelling pending bookings even if they have already "started" (:pr:`5995`)
+- Disallow switching the repeat frequency of an existing room booking from weekly to monthly
+  or vice versa (:pr:`5999`)
 
 Accessibility
 ^^^^^^^^^^^^^